pyproject.toml
[tool.poetry]
name = "phone-email-auth-api"
version = "0.1.0"
description = ""
authors = ["Mustafa S. Hassan <developer.mustafa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-decouple = "^3.8"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-allauth = {extras = ["socialaccount"], version = "^0.63.3"}
djangorestframework = "^3.15.2"
psycopg = {extras = ["binary"], version = "^3.1.19"}
twilio = "^9.2.1"
dj-rest-auth = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
